Improve sesame service test
We splitted the test in two parts :

- When the user is already logged via the cookie
- When the user is not logged

Also, we cleaned the Http verify assertions in a dedicated test.

Moreover, we added a new rule: We do not want to call sesame server when user is not logged and not have any cookie. It will prevent errors and unwanted http calls.