MM-17: (fix) Fix chrome cookie error on localhost
The bug occurs because we use a cookie on different domains:
* localhost for babel
* sesame.vidal.net for sesame

On Google Chrome, there is a rule wich specify that a cookie have to be
setted with attribute 'Secure' and 'SameSite = None' in its header to
not be rejected.

This rule creates a second problem, non-secured websites (HTTP) can not define cookies with 'Secure' attribute.
So we have to use a secured website (HTTPS).

But this rule works differently on Firefox because it is disabled on localhost.

To avoid this problem, we now use 'username' as a PathParameter instead
of CookieParam in our APIs.